+++ gax-httpjson/src/test/java/com/google/api/gax/httpjson/RetryingTest.java
@@ -27,7 +27,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.google.api.gax.grpc;
+package com.google.api.gax.httpjson;
 
 import com.google.api.core.ApiFuture;
 import com.google.api.core.ApiFutures;
